feat(customclusters): refactoring the mqtt topic
- Supporting new topic `bridge/definitions`
  Introduced in https://github.com/Koenkk/zigbee2mqtt/pull/22583
- Removing previously added files from `zigbee-herdsman`, but keeping
  the enum one
