+++ src/components/attribute-picker/index.tsx
@@ -1,10 +1,10 @@
 import React, { ChangeEvent, InputHTMLAttributes } from 'react';
 import { Attribute, Cluster, Device } from '../../types';
 
-import {DataType} from '../../zcl/definition/enums';
-import {Clusters} from '../../zcl/definition/cluster';
+import {DataType} from '../../ZCLenums';
 
 import { useTranslation } from 'react-i18next';
+import store from '../../store';
 
 export interface AttributeDefinition {
     ID: number;
@@ -19,11 +19,13 @@ interface AttributePickerProps {
     label?: string;
     onChange: (attr: Attribute, definition: AttributeDefinition) => void;
 }
+
 // eslint-disable-next-line react/prefer-stateless-function
 export default function AttributePicker(
     props: AttributePickerProps & Omit<InputHTMLAttributes<HTMLSelectElement>, 'onChange'>,
 ): JSX.Element {
     const { cluster, device, onChange, label, value, ...rest } = props;
+    const {bridgeDefinitions} = store.getState();
     const { t } = useTranslation('zigbee');
     const onChangeHandler = (e: ChangeEvent<HTMLSelectElement>): void => {
         const { value: inputValue } = e.target;
@@ -35,14 +37,14 @@ export default function AttributePicker(
             onChange(inputValue, attributeInfo);
         }
     };
-    // Retrieve Cluster attributes: from ZH first, then from device definition (expecting custom cluster as part of the definition)
+    // Retrieve Cluster attributes: from ZH first, then from device definition
     const getClusterAttributes = (cluster) : string[] | Readonly<Record<string, Readonly<AttributeDefinition>>> => {
-        const _clusterDefinition = Clusters[cluster];
+        const _clusterDefinition = bridgeDefinitions.clusters[cluster];
 
         if (_clusterDefinition) {
             return _clusterDefinition.attributes;
         } else {
-            const _customClusters = device.definition?.custom_clusters;
+            const _customClusters = bridgeDefinitions.custom_clusters[device.ieee_address];
 
             if (_customClusters && _customClusters[cluster]) {
                 return _customClusters[cluster].attributes;
