+++ src/types.ts
@@ -31,6 +31,35 @@ export type Attribute = string;
 
 export type Endpoint = string | number;
 
+export interface AttributeDefinition {
+    ID: number;
+    type: number;
+    manufacturerCode?: number;
+}
+export interface Parameter {
+    name: string;
+    type: number;
+}
+export interface CommandDefinition {
+    ID: number;
+    parameters: readonly Parameter[];
+    response?: number;
+}
+
+export interface ClusterDefinition {
+    ID: number;
+    name?: string;
+    manufacturerCode?: number;
+    attributes: Readonly<Record<string, Readonly<AttributeDefinition>>>;
+    commands: Readonly<Record<string, Readonly<CommandDefinition>>>;
+    commandsResponse: Readonly<Record<string, Readonly<CommandDefinition>>>;
+}
+
+export interface BridgeDefinitions {
+    clusters: Readonly<Record<Cluster, Readonly<ClusterDefinition>>>,
+    custom_clusters: Readonly<Record<IEEEEAddress, Readonly<CustomClusters>>>,
+}
+
 export interface Meta {
     revision: number;
     transportrev: number;
